skills/jananthan30--Resume-Builder/requirements.txt
anthropic>=0.18.0
pdfplumber>=0.10.0
python-docx>=1.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
sentence-transformers>=2.2.0
nltk>=3.8.0
textstat>=0.7.0
networkx>=3.0
rank-bm25>=0.2.2
numpy>=1.24.0
openpyxl>=3.1.0
pandas>=2.0.0
fastmcp>=3.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
PyJWT>=2.8.0
stripe>=7.0.0
python-dotenv>=1.0.0
trafilatura>=1.12.0
